Allow continuous pitch rotation in vtkInteractorStyleTerrain2::Rotate
Previously the function used the Azimuth and Elevation API of vtkCamera
which doesn't manage the view up vector.  The view up vector was fixed.
This new implementation computes the current pitch and yaw and from the
camera, applies a delta allowing for continuous rotation, and then computes
a new view up vector.  After this change, the view up vector of the
camera changes with camera rotation and so old code that assumed the up
vector was constant may need to be updated.