Improve pick tolerance code to compute tolerance based on view thickness
In older versions of VTK the pick tolerance parameter was a percentage of
the screen size in pixels, but now tolerance is related to the size of
the 3D scene.  The problem (or something related) is described here:

https://public.kitware.com/pipermail/vtkusers/2012-June/074985.html

This commit introduces a new function to compute a good tolerance using
the view thickness: the distance between the near/far clipping planes.