Refactor depthscanner.py to make the extra views optional
The depth scanner can visualize the depth image and pointcloud in separate
views.  Previously these views were always constructed, but this is
wasteful because it creates extra OpenGL contexts if you don't actually
need to see that data.  For example, when using the depth scanner in a
headless rendering environment.